Create P3WeightConverter.cpp
finished program 3 which was to convert pounds and ounces to grams and kilo grams. One important thing I found was that when converting a double to a int for positive numbers, you are essentially removing the decimal point, not rounding. To round you can use math.round(), or you can simply add 0.5 before casting.